refactored code, renamed expenses file
