+++ server/index.ts
@@ -3,7 +3,7 @@ import categoryRouter from "./routes/categories";
 import { pool } from "./db";
 import userRouter from "./routes/users";
 import cors from "cors";
-import expenseRouter from "./routes/expense";
+import expenseRouter from "./routes/expenses";
 import savedCategoriesRouter from "./routes/savedCategories";
 import weeklySummaryRouter from "./routes/weeklySummary";
 import chartRouter from "./routes/charts";
@@ -28,6 +28,6 @@ app.get("/", (req: Request, res: Response) => {
 });
 
 app.listen(port, async () => {
-  console.log("Listening to ", port);
   await pool.connect();
+  console.log("Listening to ", port);
 });
